refactor: update utilities dir structure
