+++ src/content/sketch/flocking/src/flocking.render.ts
@@ -1,8 +1,8 @@
 import { Boid } from "./boid";
 import { randInt } from "@utils/random";
-import type { Renderer } from "@components/canvas";
 import { QuadTree, Rectangle, Point, Circle } from "@utils/quad-tree";
 import { options as optionsDefault, type Options } from "./flocking.options";
+import type { Renderer } from "@components/canvas";
 import type { Sketch } from "@components/sketch/sketch.types";
 
 export const sketch: Sketch<Options> = (options = optionsDefault) => {
