+++ src/lib/utils/helpers/endcode-title-id.ts
@@ -1,0 +1,10 @@
+/**
+ *
+ */
+export function encodeTitleId(title: string) {
+	const id = title
+		.toLowerCase()
+		.replace(/\s/g, "-");
+
+	return encodeURI(id);
+}
